Add uint32_t i32 field to Value union
This allows using i32 values directly what is less confusing than
previous emulation with storing these as uint64_t.

+++ lib/fizzy/instantiate.cpp
@@ -289,7 +289,7 @@ std::unique_ptr<Instance> instantiate(std::unique_ptr<const Module> module,
     {
         // Offset is validated to be i32, but it's used in 64-bit calculation below.
         const uint64_t offset =
-            eval_constant_expression(data.offset, imported_globals, globals).i64;
+            eval_constant_expression(data.offset, imported_globals, globals).i32;
 
         if (offset + data.init.size() > memory->size())
             throw instantiate_error{"data segment is out of memory bounds"};
@@ -304,7 +304,7 @@ std::unique_ptr<Instance> instantiate(std::unique_ptr<const Module> module,
     {
         // Offset is validated to be i32, but it's used in 64-bit calculation below.
         const uint64_t offset =
-            eval_constant_expression(element.offset, imported_globals, globals).i64;
+            eval_constant_expression(element.offset, imported_globals, globals).i32;
 
         if (offset + element.init.size() > table->size())
             throw instantiate_error{"element segment is out of table bounds"};
